+++ docs/source/whatsnew.rst
@@ -4,7 +4,9 @@ What's New
 ==========
 v3.1.0
 ------
-
+- PERFORMANCE: **wxcadm** previously retrieved too much data when Orgs and Locations were initialized. The API calls have been moved so they are only performed when the relevant data is needed.
+- PERFORMANCE: :py:meth:`Location.numbers` and :py:meth:`Location.asailable_numbers` now don't pull all numbers for the Org, only the Location
+- **wxcadm** now supports Wholesale partners with the :py:class:`WholesaleCustomer` and :py:class:`WholesaleSubscriber` classes
 - :py:meth:`Location.set_announcement_language()` added to update Announcement Language for Locations, Users and Features
 - BUG FIX: Fixed exception when XSI.profile is called and no profile is returned
 - XSI calls, either :py:meth:`new_call()` or :py:meth:`originate()` now accept a ``phone`` argument to specify which phone is to be used to make the call.